Extract name printer into separate package
